# license-application-service/src/main/resources/api.yaml
openapi: 3.1.0
info:
  title: Mallorca Holiday Rental Licensing API
  version: 1.0.0
  description: |
    API for managing applications, licenses, and payments for the Mallorca Holiday Rental Licensing system.
    Supports GDPR-compliant user management through Keycloak integration and ensures transparent, fair allocation of ETV licenses.
servers:
  - url: https://localhost:8080/
...
security:
  - KeycloakOAuth2:
      - openid
      - profile
      - email
paths:
...
  /ballot-periods:
    get:
      tags: [Ballot Periods]
      summary: List all ballot periods
      description: Retrieve a list of all ballot periods, including their start and end dates.
      operationId: listBallotPeriods
      responses:
        "200":
          description: List of ballot periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BallotPeriodResource"
        "401":
          description: Unauthorized
    post:
      tags: [Ballot Periods]
      summary: Create a ballot period
      description: Create a new ballot period by defining its start and end dates.
      operationId: createBallotPeriod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Ballot period created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BallotPeriodResource"
        "409":
          description: Conflict - overlapping ballot periods are not allowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error_code:
                    type: string
                    example: "OVERLAPPING_PERIOD"
                  message:
                    type: string
                    example: "The specified start and end dates overlap with an existing ballot period."
  /ballot-periods/current:
    get:
      tags: [Ballot Periods]
      summary: Get date of most current
      description: >
        Retrieve the start and end dates of the current,
        next (if none ongoing) or last ballot period (if none ongoing or upcoming).
      operationId: getBallotPeriod
      responses:
        "200":
          description: Current, upcoming or last ballot period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentBallotPeriodResource"
        "204":
          description: No active ballot periods found
  /ballot-periods/{period_id}:
    get:
      tags: [Ballot Periods]
      summary: Retrieve ballot period details
      description: >
        Retrieves the details and configuration for a specific ballot period, 
        including its start and end dates and other metadata.
      operationId: getBallotPeriodDetails
      parameters:
        - in: path
          name: period_id
          required: true
          schema:
            type: integer
          description: Unique identifier for the ballot period
      responses:
        "200":
          description: Ballot period details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BallotPeriodResource"
  /ballot-periods/{period_id}/entries:
    get:
      tags: [Ballot Periods]
      summary: List entries for a ballot period
      description: List all submitted applications/entries for a specific ballot period.
      operationId: getBallotPeriodEntries
      parameters:
        - in: path
          name: period_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of entries for the period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationResource"
  /ballot-periods/{period_id}/lottery:
    post:
      tags: [Ballot Periods]
      summary: Run the lottery round for a ballot period
      description: >
        Triggers the allocation (lottery) process for this ballot period,
        assigning selected status as true/false to eligible applications within the given period.
        It is important to note, that a period has to be over.
      operationId: runLotteryForBallotPeriod
      parameters:
        - in: path
          name: period_id
          required: true
          schema:
            type: integer
        - name: licensesCount
          description: Number of licenses to allocate in this lottery round
          in: query
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                license_type:
                  $ref: "#/components/schemas/LicenseTypeApiEnum"
      responses:
        "200":
          description: Lottery results with updated statuses for applications in the period
          content:
            application/json:
              schema:
                type: object
                properties:
                  selected_applications:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApplicationResource"
                  not_selected_applications:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApplicationResource"
        "400":
          description: Bad request - invalid input data. Can be caused because period is not over yet
...
components:
  securitySchemes:
    KeycloakOAuth2:
      type: oauth2
      description: OAuth2 with Keycloak JWT Bearer tokens.
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: Profile information access
            email: Email access
  schemas:
...
    ApplicationResource:
      type: object
      description: Represents a user's request for a holiday rental license.
      properties:
        id:
          type: integer
          description: Unique application identifier.
          example: 201
        user_id:
          type: string
          format: uuid
          description: ID of the applicant user.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        license_type:
          $ref: "#/components/schemas/LicenseTypeApiEnum"
        cadastral_reference:
          type: string
          minLength: 20
          maxLength: 20
          description: Unique cadastral code (property reference).
          example: 1234567AB9999C0001DE
        applied_at:
          type: string
          format: date-time
          description: When the application was submitted.
          example: 2025-05-20T08:00:00Z
        changed_at:
          type: string
          format: date-time
          description: When the application was last updated.
          example: 2025-06-01T10:00:00Z
        application_status:
          $ref: "#/components/schemas/ApplicationStatusApiEnum"
        remarks:
          type: string
          description: Additional information or comments.
          example: Needs urgent review.
      required:
        - id
        - user_id
        - license_type
        - cadastral_reference
        - application_status
...
    ApplicationStatusApiEnum:
      type: string
      description: Status values for an application.
      enum:
        - DRAFT # Saved but not submitted
        - DOCUMENTS_SUBMITTED # Documents uploaded
        - VERIFICATION_PENDING # Awaiting verification
        - AWAITING_PAYMENT # Waiting for user payment
        - PAYMENT_RECEIVED # Payment details received
        - SUBMITTED # Application formally submitted
        - IN_BALLOT # Entered into selection/ballot
        - SELECTED # Chosen/approved via ballot
        - NOT_SELECTED # Not chosen in ballot
        - UNDER_REVIEW # In manual review process
        - APPROVED # License approved
        - REJECTED # Application rejected or payment denied
        - EXPIRED # Application expired (not continued)
        - CANCELLED # Withdrawn/cancelled by user
...
    LicenseTypeApiEnum:
      type: string
      description: License categories for holiday rentals.
      enum:
        - ETV
        - ETVPL
        - ETV60
...
    BallotPeriodResource:
      type: object
      properties:
        ballot_period_id:
          type: integer
          description: Unique identifier for the ballot period
        start_date:
          type: string
          format: date-time
          description: Ballot period start, ISO 8601 format
        end_date:
          type: string
          format: date-time
          description: Ballot period end, ISO 8601 format
        totalApplications:
          type: integer
          description: Total number of applications received in this ballot period
    CurrentBallotPeriodResource:
      type: object
      properties:
        start_date:
          type: string
          format: date-time
          description: Ballot period start, ISO 8601 format
        end_date:
          type: string
          format: date-time
          description: Ballot period end, ISO 8601 format
...